--- src/constants/channel-constants.js
@@ -1,12 +1,0 @@
-'use strict';
-
-module.exports = {
-  CHANNEL_LAUNCH_UUID_V4_BROWSER_WINDOW: 'CHANNEL_LAUNCH_UUID_V4_BROWSER_WINDOW',
-  CHANNEL_LAUNCH_UUID_V5_BROWSER_WINDOW: 'CHANNEL_LAUNCH_UUID_V5_BROWSER_WINDOW',
-  CHANNEL_LAUNCH_JWT_DECODER_BROWSER_WINDOW: 'CHANNEL_LAUNCH_JWT_DECODER_BROWSER_WINDOW',
-  CHANNEL_LAUNCH_BASE64_ENCODER_DECODER_DECODER_BROWSER_WINDOW:
-    'CHANNEL_LAUNCH_BASE64_ENCODER_DECODER_DECODER_BROWSER_WINDOW',
-  CHANNEL_LAUNCH_EPOCH_BROWSER_WINDOW: 'CHANNEL_LAUNCH_EPOCH_BROWSER_WINDOW',
-  CHANNEL_LAUNCH_JSON_FORMATTER_BROWSER_WINDOW: 'CHANNEL_LAUNCH_JSON_FORMATTER_BROWSER_WINDOW',
-  CHANNEL_LAUNCH_XML_FORMATTER_BROWSER_WINDOW: 'CHANNEL_LAUNCH_XML_FORMATTER_BROWSER_WINDOW'
-};
