--- src/handlers/ipc/index.js
@@ -1,8 +1,0 @@
-'use strict';
-const mainIPCHandler = require('./main-ipc');
-
-function ipcHandler({ appBrowserWindows }) {
-  mainIPCHandler({ appBrowserWindows });
-}
-
-module.exports = ipcHandler;
