+++ src/main-process/main.js
@@ -1,25 +1,60 @@
-'use strict';
-const { app } = require('electron');
-const applicationMenu = require('./application-menu');
-const appBrowserWindows = require('./browser-windows')();
-const ipcHandler = require('../handlers/ipc');
+"use strict";
+const { app, BrowserWindow } = require("electron");
+const windowStateKeeper = require("electron-window-state");
+const applicationMenu = require("./application-menu");
 
-app.whenReady().then(async () => {
-  await applicationMenu();
-  await appBrowserWindows.mainBrowserWindow.createWindow();
-  ipcHandler({ appBrowserWindows });
-});
+// global variable to prevent it from getting garbage collected.
+let mainWindow;
+
+function createWindow() {
+  const windowState = windowStateKeeper({
+    defaultWidth: 800,
+    defaultHeight: 600
+  });
+
+  mainWindow = new BrowserWindow({
+    x: windowState.x,
+    y: windowState.y,
+    width: windowState.width,
+    height: windowState.height,
+    minWidth: 800,
+    minHeight: 500,
+    backgroundColor: '#25282c',
+    webPreferences: {
+      contextIsolation: false,
+      nodeIntegration: true
+    }
+  });
+
+  // for dev work
+  mainWindow.webContents.openDevTools()
+
+  windowState.manage(mainWindow);
+
+  mainWindow.loadFile(__dirname + '/../static/app.html');
+
+  mainWindow.on('closed', () => {
+    mainWindow = null;
+  });
+}
+
+app
+  .whenReady()
+  .then(() => {
+    applicationMenu();
+    createWindow();
+  });
 
 // Quit the app when all windows are closed. Not for macOS.
-app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') {
+app.on("window-all-closed", () => {
+  if (process.platform !== "darwin") {
     app.quit();
   }
 });
 
 // When app icon is clicked and app is running then recreate BrowserWindow (macOS)
-app.on('activate', async () => {
-  if (!appBrowserWindows.mainBrowserWindow.getBrowserWindowInstance()) {
-    await appBrowserWindows.mainBrowserWindow.createWindow();
+app.on("activate", async () => {
+  if (!mainWindow) {
+    createWindow();
   }
 });
