+++ src/static/css/app.css
@@ -1,0 +1,3 @@
+body {
+    -webkit-user-select: none;
+}
